Ability to set write deadline, and bugfixes
If the connection is dropped and a message is in the process of being published, it will block and
prevent the ping handling from realizing that the connection has been lost. By allowing the user
to set a write deadline, they specify the max time we should block until we error on the publish
with a timeout.

Signed-off-by: Christie Deiters <cdeiters@lutron.com>