Bugfix: c.stop closed twice in some startup conditions.
Sometimes net.connect sees an issue and pushes an error on c.errors before returning an rc of CONN_FAILURE. It is also possible
to exit net.connect with a non-successful rc but NOT get an error on c.errors. So I updated net.connect to not push errors on
c.errors, but instead just return CONN_FAILURE and allow client.Start to put the erorr on the channel. In addition, I made c.errors
a buffer size 1 and updated all pushes to that channel to be a select with an empty default. This allows us to use c.errors as a flag channel so that
the first error can always get through, but subsequent attempts to push errors will not block forever.

Also, the begin chan ConnRC in MqttClient was not longer being initialized so I removed it. I also removed the call to close it in net.connect, since
closing a nil channel was causing a crash.

Change-Id: I0988ad51e7740b03fff832084ea65dcbce173884
Signed-off-by: Christie Deiters <cdeiters@lutron.com>