Add tokens and move message ids
Tokens are a way to get the result of an operation, they are provided
for Publish, Subscribe, SubscribeMultiple and Unsubscribe. For Publish
and Unsubscribe they allow you to wait for the operation to complete,
this indicates that the QoS flow for that message has finished or the
server has responded with an UNSUBACK.
For Unsubscribe the token includes a map of the topics the client
requested to subscribe to along with the QoS values, or error codes
from the server for each topic.
Message ids have moved to become an embedded property of the Client
struct rather than part of options. Message ids are now also a map
of uint16 to token.