Automatic reconnect and directly editable client options
When the client detects that it has lost its connection to the server
it will automatically try and reconnect with the same settings as
before. The connection lost callback is still called with an error
indicating why the connection was lost but you are not expected to
take any actions to reconnect, there is an internal reconnect function
that will back off the time between attempted reconnect exponentially
from 1 second up to the MaxReconnectInterval in the client options.
On initial connect and any subsequent reconnection the OnConnect
callback will be called, in the case of non durable session this may be
used to do subscription actions.

It is now possible to directly assign values in ClientOptions, the
assignment functions are still provided for convenience and backward
compatibility.