Bugfixes: Cleaning up Go-Routines Left Over in case of Bad Username and Password during login.
Adding ability to stop the messageid object when a connection is torn down, releasing the go-routine. Also, adding ability to stop the outgoing channel when no message is being sent (in the
case of error during initial connection, like bad username/password). Added a timeout to all attempts to send on outgoing via oboung and oboundP, to prevent them blocking forever if outgoing
is no longer running.

Change-Id: I0f9c3b4154fdeda1d552f8e7c9c14a6f7620b368
Signed-off-by: Christie Deiters <cdeiters@lutron.com>