Mark receiveTimer as read when triggered
We were not marking this time as read so in the case
where the client was sending only QoS0 messages and
not receiving anything at all from the server the timer
would trigger to send a pingreq, in the sendping function
we would call reset on this timer, which calls stop
in stop() t.Timer.Stop() is false, the timer has expired
and t.readfrom is false, even though we have actually
consumed the value. Line 80 overall evaluates to true
and we block on trying to drain the timer channel.
The client gets disconnected for a ping timeout (ironically)
but the autoreconnect cannot kick in because the ping
goroutine wont exit because it's blocked on a channel
read that will never come in.

resolves #92