Add support for setting the Mosquitto bridge-connection option
Mosquitto (and some other MQTT brokers) support a non-standard
option of setting the high bit of the protocol version to indicate
to the broker that this client connection is bridge-style.

In particular, this allows the broker to suppress sending a client
messages that it published.

We are using the paho.mqtt.golang client library to create a bridge
application for bridging Mosquitto to other non-MQTT messaging servers
and have updated the client library to support protocol versions with
the 0x80 bit set to enable this option.

The method of implimentation submitted in this version of the pull request
is the one that makes the most minimal change to paho.mqtt.golan while
still allowing access to this functionality and does not introduce any
new API functions.

An alternate version of this enhancement has been considered, that might
be more consistent with the current API structure, but introduces a new API
call would be:
    func (o *ClientOptions) SetBridgeConnection(enable bool) *ClientOptions

This option would then trigger merging of the bridge bit with the protocol
version when constructing the connection message wire payload.

A version like this would also be acceptable for our use case.

Signed-off-by: Dave Miller <dmiller@tibco.com>