AddBroker: Escape % in Addresses
The url.Parse(server) will fail if given an address containing a
non escaped % which is the case for an IPv6 address with a zone
indication.

Ex.
[fe80::a604:5838:8d8a:e532%wlp58s0]:1883 => [fe80::a604:5838:8d8a:e532%25wlp58s0]:1883

Regexp explanation:
`%(25)?`
% - Match % literally (case sensitive)
(25)? -  Match 25 literally between zero and one time.

=> Will match any % followed by 25 or not.
The Regexp.ReplaceAll(...) will replace all matches of the regular
expression above and replace them with %25. Which is the URL escape code
for %.

Signed-off-by: Måns Ansgariusson <mansgariusson@gmail.com>